Improve Paywall view controller by introducing sharing on FB/Twitter. Introduce CNSalesPitch object which serves as a container for all the dynamic pitches/freebie target information that comes from server.
Due to label sizing issues (you have to know exact size of text
to place it precisely within frame, i.e. label.frame = sizeoftext),
decided to get rid of manual approach for setting frames and use
autolayout.

it turned out to be a bit less of a pain for custom stuff and is a
bit less of manual work for 4 screen sizes of iPhone.